Move get_ip into accountlockout project
Move get_ip and update usages in the utils.py file.
