+++ accountlockout/config.py
@@ -1,0 +1,8 @@
+DEFENDER_LOGIN_FAILURE_LIMIT = 5
+DEFENDER_BEHIND_REVERSE_PROXY = False
+DEFENDER_DISABLE_IP_LOCKOUT = True
+DEFENDER_COOLOFF_TIME = 600
+DEFENDER_LOCKOUT_TEMPLATE = 'locked_out.html'
+DEFENDER_REDIS_URL = 'redis://cache:6379/0'
+DEFENDER_ACCESS_ATTEMPT_EXPIRATION =24
+DEFENDER_USERNAME_FORM_FIELD = 'auth-username'
